# laravel/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "TKK Control API",
        "description": "API documentation for TKK-Control",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "API server"
        }
    ],
    "paths": {
...
        "/api/controls": {
            "get": {
                "tags": [
                    "Controls"
                ],
                "summary": "Get all controls",
                "operationId": "c77f525a2a1446577d038002698a65bf",
                "responses": {
                    "200": {
                        "description": "List of controls",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Controls"
                ],
                "summary": "Create a new control",
                "operationId": "049b68992d9cbcb1ba1db2510ab39199",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "km_point"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "CP1"
                                    },
                                    "km_point": {
                                        "type": "number",
                                        "example": 12.5
                                    },
                                    "responsible": {
                                        "type": "string",
                                        "example": "John Doe"
                                    },
                                    "phone": {
                                        "type": "string",
                                        "example": "+34123456789"
                                    },
                                    "status": {
                                        "type": "string",
                                        "enum": [
                                            "preparing",
                                            "open_requested",
                                            "opened",
                                            "close_requested",
                                            "closed"
                                        ],
                                        "example": "preparing"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Control created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Validation error"
                    }
                }
            }
        },
        "/api/controls/{id}": {
            "get": {
                "tags": [
                    "Controls"
                ],
                "summary": "Get a control by ID",
                "operationId": "814fb0a47f987c12d3e2f7b9f0612017",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Control ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Control details with relations",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Control not found"
                    }
                }
            },
            "patch": {
                "tags": [
                    "Controls"
                ],
                "summary": "Update a control status",
                "description": "Update the status of a control with allowed state transitions.",
                "operationId": "3014109f4371432ca4cffeab11e7c709",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Control ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "status"
                                ],
                                "properties": {
                                    "status": {
                                        "type": "string",
                                        "enum": [
                                            "preparing",
                                            "open_requested",
                                            "opened",
                                            "close_requested",
                                            "closed"
                                        ],
                                        "example": "open_requested"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Status updated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid state transition"
                    },
                    "404": {
                        "description": "Control not found"
                    }
                }
            }
        },
        "/api/admin/controls/import": {
            "post": {
                "tags": [
                    "Controls"
                ],
                "summary": "Import controls from CSV",
                "description": "Upload a CSV file to import multiple controls.",
                "operationId": "de1b593aef4a200686c9c4798fe7aef0",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "csv"
                                ],
                                "properties": {
                                    "csv": {
                                        "description": "CSV file with controls",
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Import result with added count and errors",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "added": {
                                            "type": "integer",
                                            "example": 5
                                        },
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid file"
                    }
                }
            }
        },
...
    },
...
}